Show multi pics.

+++ app/src/main/java/com/christmas/strawberryweibo/model/entity/StatusListWrapper.java
@@ -5,8 +5,8 @@ import com.google.gson.annotations.SerializedName;
 import java.util.List;
 
 public class StatusListWrapper {
-  @SerializedName("previous_cursor") public int previousCursor;
-  @SerializedName("total_number") public int totalNumber;
+  @SerializedName("previous_cursor") public long previousCursor;
+  @SerializedName("total_number") public long totalNumber;
   @SerializedName("next_cursor") public long nextCursor;
 
   public List<Status> statuses;
